+++ R/warbleR-package.R
@@ -38,6 +38,10 @@
 #'   
 #' @section Managing sound files:
 #'   
+#'   \code{\link{read_wave}}: Read wave files into 'wave' objects
+#'   
+#'   \code{\link{read_sound_file}}: Read sound files into 'wave' objects
+#'   
 #'   \code{\link{selection_table}}: Create 'selection_table' class objects
 #'   
 #'   \code{\link{mp32wav}}: Convert several .mp3 files in working directory to .wav