+++ NEWS.md
@@ -1,13 +1,20 @@
 # *warbleR 1.1.25*
 
+## Bug fixes:
+
+* Negative gaps values in `song_param()`
 
 ## New functions:
 
+* 'optimize_autodetec()' for tunning detection parameters in 'autodetec()'
 * 'envelope()' a C-compiled version of 'seewave::env()'
 * new function 'find_peaks()' to detect signals on sound files using 'xcorr()' output
+* 'read_sound_file()' to read wav, wav and mp3 files as well as wave objects in extended selection tables
 
 ## Changes and additions:
 
+* Remove comment columns in 'lbh_selec_table'
+* Selection table printing method uses huxtable package for text coloring
 * 'suffix' argument added to `lspec()`
 * `autodetec()` does not create image files with spectrograms any more
 * new arguments added to `autodetec()`: 'hold.time' for merging selections, 'thinning' for decreasing envelope size